bot/requirements.txt
aiogram>=2.25.1,<3.0.0
python-dotenv>=1.0.0